feat(applications): migrate new application form to server/client architecture
- Split new application form into server component (metadata, navigation) and client component (form state, submission)
- Add new applications API route POST endpoints
- Maintain exact same functionality while improving performance and SEO
- Follow Next.js 14 App Router best practices with proper component separation
